add float support

+++ calc.py
@@ -1,7 +1,7 @@
 # This is a code for addition and subtraction
 def compute(expression):
     num0, operator, num1 = expression.split(' ')
-    num0, num1 = int(num0), int(num1)
+    num0, num1 = float(num0), float(num1)
     if operator == '+':
         return num0 + num1
     elif operator == '-':
